refactored GameStateLogic into serverData
